change sms: call to imessage://
``sms:`` is meant strictly for phone numbers and seems to force enable MMS when used on email addresses (assuming it even works in the first place). however, the ``imessage://`` protocol (strictly for Apple devices) seems to call up Apple Messages and start a new iMessage thread just fine on iOS 17.4 Dev Beta. this will require some testing before it gets pulled, but it seems to be a good enough replacement.